Cleaned up Save/Load System
Save/Load system from main menu completed and cleaned.
