+++ Assets/Scripts/Managers/ButtonManager.cs
@@ -1,18 +1,9 @@
-using UnityEditor.SceneManagement;
 using UnityEngine;
-using UnityEngine.EventSystems;
 using UnityEngine.SceneManagement;
-using System;
 namespace Managers
 {
 
     public class ButtonManager : MonoBehaviour {
-        int n;
-
-        public void OnButtonPress(){
-            n++;
-            Debug.Log("Button clicked " + n + " times.");
-        }
         public void RestartGame()
         {
             Scene scene = SceneManager.GetActiveScene(); 
@@ -34,29 +25,13 @@ namespace Managers
         {
             SceneManager.LoadScene("SampleScene");
             GameStartManager.PlayingSavedGame = true;
-            Debug.Log("Teeeessssststtt");
-           // PlayerSaveData data = SaveSystem.LoadPlayer();
-           // PlayerController playerInstance = new PlayerController();
-           // if (playerInstance != null)
-          //  {
-            //    Debug.Log("player:" + playerInstance.transform.position);
-
-               // data.ApplyPlayerSavedData(playerInstance);
-          //  }
         }
         
 
         public void SaveGame(GameObject savedGameIndicator)
         {
-
-            //PlayerController.instance.saved = true;
             SaveSystem.SavePlayer(PlayerController.instance);
-            //Debug.Log("Player controller data pos x:"+PlayerController.instance.transform.position.x+" y:"
-           //           +PlayerController.instance.transform.position.y+" z:"+PlayerController.instance.transform.position.z
-            //          +" currenthealth:"+PlayerController.instance.currentHealth+" mana:"+PlayerManager.instance.currentMana);
-
             SavedGameIndicator.Create(transform.position, savedGameIndicator);
-
         }
     }
 }
