+++ Assets/Scripts/Managers/InteractableManager.cs
@@ -1,8 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.Experimental.TerrainAPI;
 
 public class InteractableManager : MonoBehaviour
 {
