--- Assets/Scripts/PlayerManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c4701d970f5941d47bca64d3371da8e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
